initial ai-splitting of configuration classes
